Ask for the address on the docs pages too
The launch-list form was one client component on the home page. It is now <NewsletterForm>, shared with a docs footer CTA that renders under every docs page — a guide reader is as good an address as a component reader, and it was the one surface with no ask on it at all. The home band drops "use client" as a side effect: nothing left in it holds state.
